Move createTaskForm hook to a context
Since we need the form logic in different places of the page, a provider
will be use to avoid prop drilling and sending the required parts of the
form logic to where it is needed
